pyproject.toml
[tool.poetry]
name = "ML-Pipeline-of-Paper-Category-Classification"
version = "0.1.0"
description = ""
authors = ["nomnomnonono <nozomu.koujiguchi1204@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12.0"
arxiv = "^1.4.8"
pandas = "^2.0.3"
python-dotenv = "^1.0.0"
kfp = "^2.1.2"
google-cloud-storage = "^2.10.0"

[tool.poetry.dev-dependencies]
pysen = {version = "^0.10.3", extras = ["lint"]}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.10.3"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = false
mypy_preset = "strict"
line_length = 88
py_version = "py39"

[[tool.pysen.lint.mypy_targets]]
paths = ["."]
